Functioning of the site

+++ src/Controller/TrickController.php
@@ -65,7 +65,7 @@ class TrickController extends AbstractController {
                 $featuredFile->move($this->getParameter("featured_directory"), $featuredFileName);
                 $featuredPath = "assets/images/tricks/featured/".$featuredFileName;
             } else {
-                $featuredPath = "assets/images/tricks/placeholder/trick_placeholder.webp";
+                $featuredPath = "assets/images/tricks/featured/placeholder/trick_placeholder.webp";
             }
 
             $media = new Media();
@@ -112,24 +112,29 @@ class TrickController extends AbstractController {
             if($mediasEmbed) {
                 $mediasEmbedList = explode("\n", $mediasEmbed);
                 foreach($mediasEmbedList as $mediaEmbedItem) {
-                    if(parse_url($mediaEmbedItem)["host"] === "www.youtube.com") {   
-                        $embedVideo = rtrim(str_replace("v=", "", parse_url($mediaEmbedItem)["query"]), "_");
-                        $embedUrl = "https://youtube.com/embed/".$embedVideo;
-                    } else if(parse_url($mediaEmbedItem)["host"] === "www.dailymotion.com") {
-                        $embedVideo = str_replace("/video/", "",  parse_url($mediaEmbedItem)["path"]);
-                        $embedUrl = "https://dailymotion.com/embed/video/".$embedVideo;
-                    } else {
+                    if(!array_key_exists("host", parse_url($mediaEmbedItem))) {
                         $this->addFlash("warning", "Veuillez ne prendre un lien que de Youtube et de Dailymotion.");
                         return $this->redirectToRoute("trick_create");
+                    } else {
+                        if(parse_url($mediaEmbedItem)["host"] === "www.youtube.com") {   
+                            $embedVideo = rtrim(str_replace("v=", "", parse_url($mediaEmbedItem)["query"]), "_");
+                            $embedUrl = "https://youtube.com/embed/".$embedVideo;
+                        } else if(parse_url($mediaEmbedItem)["host"] === "www.dailymotion.com") {
+                            $embedVideo = str_replace("/video/", "",  parse_url($mediaEmbedItem)["path"]);
+                            $embedUrl = "https://dailymotion.com/embed/video/".$embedVideo;
+                        } else {
+                            $this->addFlash("warning", "Veuillez ne prendre un lien que de Youtube et de Dailymotion.");
+                            return $this->redirectToRoute("trick_create");
+                        }
+
+                        $media = new Media();
+                        $media->setIdTrick($trickId);
+                        $media->setType("embed");
+                        $media->setSrc($embedUrl);
+                        $media->setFeatured(false);
+
+                        $entityManager->persist($media);
                     }
-
-                    $media = new Media();
-                    $media->setIdTrick($trickId);
-                    $media->setType("embed");
-                    $media->setSrc($embedUrl);
-                    $media->setFeatured(false);
-
-                    $entityManager->persist($media);
                 }
             }
 
@@ -213,15 +218,9 @@ class TrickController extends AbstractController {
             return $this->redirectToRoute("home");
         }
 
-        $trick = $trickRepository->findOneBy(["slug" => $trickSlug]);
-        $medias = $mediaRepository->findBy(["idTrick" => $trick->getId()]);
-        $trickGroups = $trickGroupRepository->findAll();
-
-        $category = $trickGroupRepository->findOneBy(["id" => $trick->getIdTrickGroup()]);
-        $featured = $mediaRepository->findOneBy(["idTrick" => $trick->getId(), "featured" => true]);
-        $embed = $mediaRepository->findBy(["idTrick" => $trick->getId(), "type" => "embed"]);
-
         if ($request->isMethod("POST")) {
+            $trick = $trickRepository->findOneBy(["slug" => $trickSlug]);
+
             $featuredForm = $request->files->get("featured_media");
             $nameForm = $request->get("name");
             $slugForm = strtolower(str_replace(" ", "-", $nameForm));
@@ -248,8 +247,7 @@ class TrickController extends AbstractController {
             
             $trick->setName($nameForm);
             $trick->setDescription($descriptionForm);
-            $trick->setIdTrickGroup($selectedGroup);
-            $trick->setSlug($slugForm);            
+            $trick->setIdTrickGroup($selectedGroup);          
             $trick->setUpdatedAt(\DateTime::createFromFormat("Y-m-d H:i:s", date("Y-m-d H:i:s")));
         
             $trickId = $trickRepository->findOneBy(["id" => $trick->getId()]);
@@ -285,31 +283,45 @@ class TrickController extends AbstractController {
             if($embedForm) {
                 $mediasEmbedList = explode("\n", $embedForm);
                 foreach($mediasEmbedList as $mediaEmbedItem) {
-                    if(parse_url($mediaEmbedItem)["host"] === "www.youtube.com") {   
-                        $embedVideo = rtrim(str_replace("v=", "", parse_url($mediaEmbedItem)["query"]), "_");
-                        $embedUrl = "https://youtube.com/embed/".$embedVideo;
-                    } else if(parse_url($mediaEmbedItem)["host"] === "www.dailymotion.com") {
-                        $embedVideo = str_replace("/video/", "",  parse_url($mediaEmbedItem)["path"]);
-                        $embedUrl = "https://dailymotion.com/embed/video/".$embedVideo;
-                    } else {
+                    if(!array_key_exists("host", parse_url($mediaEmbedItem))) {
                         $this->addFlash("warning", "Veuillez ne prendre un lien que de Youtube et de Dailymotion.");
                         return $this->redirectToRoute("trick_edit", ["trickSlug" => $trick->getSlug()]);
-                    }
+                    } else {
+                        if(parse_url($mediaEmbedItem)["host"] === "www.youtube.com") {   
+                            $embedVideo = rtrim(str_replace("v=", "", parse_url($mediaEmbedItem)["query"]), "_");
+                            $embedUrl = "https://youtube.com/embed/".$embedVideo;
+                        } else if(parse_url($mediaEmbedItem)["host"] === "www.dailymotion.com") {
+                            $embedVideo = str_replace("/video/", "",  parse_url($mediaEmbedItem)["path"]);
+                            $embedUrl = "https://dailymotion.com/embed/video/".$embedVideo;
+                        } else {
+                            $this->addFlash("warning", "Veuillez ne prendre un lien que de Youtube et de Dailymotion.");
+                            return $this->redirectToRoute("trick_edit", ["trickSlug" => $trick->getSlug()]);
+                        }
 
-                    $media = new Media();
-                    $media->setIdTrick($trickId);
-                    $media->setType("embed");
-                    $media->setSrc($embedUrl);
-                    $media->setFeatured(false);
-                    $entityManager->persist($media);
+                        $media = new Media();
+                        $media->setIdTrick($trickId);
+                        $media->setType("embed");
+                        $media->setSrc($embedUrl);
+                        $media->setFeatured(false);
+                        $entityManager->persist($media);
+                    }
                 }
             }
 
+            $trick->setSlug($slugForm);  
             $entityManager->flush();
 
             $this->addFlash("success", "Le trick a été modifié !");
             return $this->redirectToRoute("trick_presentation", ["trickSlug" => $trick->getSlug()]);
         }    
+
+        $trick = $trickRepository->findOneBy(["slug" => $trickSlug]);
+        $medias = $mediaRepository->findBy(["idTrick" => $trick->getId()]);
+        $trickGroups = $trickGroupRepository->findAll();
+
+        $category = $trickGroupRepository->findOneBy(["id" => $trick->getIdTrickGroup()]);
+        $featured = $mediaRepository->findOneBy(["idTrick" => $trick->getId(), "featured" => true]);
+        $embed = $mediaRepository->findBy(["idTrick" => $trick->getId(), "type" => "embed"]);
         
         $data = [
             "name" => $trick->getName(),
@@ -329,7 +341,7 @@ class TrickController extends AbstractController {
 
 
     #[Route(name: "trick_delete", path: "/trick/delete/{trickSlug}")]
-    public function delete(TrickRepository $trickRepository, string $trickSlug, MediaRepository $mediaRepository, EntityManagerInterface $entityManager): Response {        
+    public function delete(CommentRepository $commentRepository, TrickRepository $trickRepository, string $trickSlug, MediaRepository $mediaRepository, EntityManagerInterface $entityManager): Response {        
         if (!$this->getUser()) {
             return $this->redirectToRoute("home");
         }
@@ -338,11 +350,18 @@ class TrickController extends AbstractController {
 
         if($trick) {
             $medias = $mediaRepository->findBy(["idTrick" => $trick->getId()]);
+            $comments = $commentRepository->findBy(["idTrick" => $trick->getId()]);
 
             foreach($medias as $media) {
                 $mediaPath = $media->getPath();
-                unlink($mediaPath);
+                if($media->getType() !== "embed" && $mediaPath !== "assets/images/tricks/featured/placeholder/trick_placeholder.webp") {
+                    unlink($mediaPath);
+                }
                 $entityManager->remove($media);
+            }
+
+            foreach($comments as $comment) {
+                $entityManager->remove($comment);
             }
 
             $entityManager->remove($trick);
@@ -394,7 +413,9 @@ class TrickController extends AbstractController {
 
         if($media) {
             $mediaPath = $media->getPath();
-            unlink($mediaPath);
+            if($media->getType() !== "embed" && $mediaPath !== "assets/images/tricks/featured/placeholder/trick_placeholder.webp") {
+                unlink($mediaPath);
+            }
             $entityManager->remove($media);
             $entityManager->flush();
